no errors without strict; restore some arrow-loading
