+++ src/glsl-read-float.d.ts
@@ -1,0 +1,5 @@
+// glsl-read-float.d.ts
+declare module 'glsl-read-float' {
+  const unpackFloat: (r: number, g: number, b: number, a: number) => number;
+  export default unpackFloat;
+}
